requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
django-modeltranslation==0.17.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
easy-thumbnails==2.7.1
Markdown==3.3.4
Pillow==8.3.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
